pyproject.toml
[project]
name = "obsidian-assintant-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "groq>=0.30.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
    "tiktoken>=0.7.0",
    "spacy>=3.7.0",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.3.0",
]
